refactor(checker): extract satisfies TS1360 builder to keep assignability under 2000 LOC
assignability.rs sat exactly at the 2000-line architecture ceiling; the
TS1360 elaboration-chain fix pushed it over. Move the self-contained
`error_type_does_not_satisfy_the_expected_type` builder into a focused
sibling module `assignability_satisfies.rs` (behavior-preserving) so the
file drops to 1914 lines with headroom. No logic change beyond the
relocation; the satisfies replace-vs-wrap fix is unchanged.
