+++ crates/tsz-checker/src/error_reporter/mod.rs
@@ -21,6 +21,7 @@ mod assignability_keyof_alias_display;
 mod assignability_literal_display;
 mod assignability_normalized_union;
 mod assignability_numeric_display;
+mod assignability_satisfies;
 mod assignability_type_helpers;
 mod call_errors;
 mod call_errors_anchors;
